LayoutTests/imported/w3c/web-platform-tests/svg/extensibility/foreignObject/containing-block-expected.txt


FAIL position assert_equals: offsetTop expected 5 but got 0
